Fix signed/unsigned comparison warning
git-svn-id: svn://scribus.net/trunk/Scribus@23780 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/colormgmt/sclcms2colorprofileimpl.cpp
@@ -113,7 +113,7 @@ bool ScLcms2ColorProfileImpl::save(QByteArray& profileData) const
 
 	// Allocate array for profile data
 	profileData.resize(bytesNeeded);
-	if (profileData.size() != bytesNeeded)
+	if (profileData.size() != static_cast<int>(bytesNeeded))
 		return false;
 	done = cmsSaveProfileToMem(m_profileHandle, profileData.data(), &bytesNeeded);
